requirements.txt
pytorch-lightning==0.8.5
torchvision
pandas==1.1.4
scikit-learn==0.23.2
opencv-python
torchsummary==1.5.1
pycm==3.0
albumentations==0.5.0
scikit-video
configargparse
ConfigArgParse==1.5.1
